requirements.txt
Django==1.11.13
django-modeltranslation==0.12.2
pytz==2018.4
